backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
 
 # Core Django packages
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
#pip install python-decouple
#pip install requests
#pip install Pillow
# Authentication & Security
djangorestframework-simplejwt==5.5.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Database & Caching
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# File handling & Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7

# Monitoring & Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0